Add NER Category and Position classes and tests
A Category is just an Enumeration which holds the different types of categories
for Named Entities found by the annotators. Currently holds a small set of
possible categories, and will probably grow in a future to hold more as needed.
It could had also be done with a sealed trait/abstract class and a set of case
objects that extend that, but by using an Enumeration, the really useful
`withName` method is already provided. The inconvenience is that Category is an
object, and thus does not define a type, so an `import Category._` is needed
wherever it will be used.

A Position is just a Value Class over an Integer that helps to reason about the
code when using it to mark the position of a Named Entity found in a Document.
It checks that the position is a positive integer, and extends the Comparable
Trait to allow easy comparison between positions (equality and hashCode are
given by Int because it is a Value Class, aka extends AnyVal).