Redesign directory/package structure
Previous package structure did not resemble the system needs, so it is
restructured to bes suit them. Mainly, classes that were in *.model.ner are now
in *.model.entity, because those three were just Entity classes (Category,
Position and Sentence).

This structure poses a better organization of the different classes that
conform the system according to what responsibilities them have. It is not
fully stabilized, so it's actually open to future rearrangements.