Add Document entity class and tests
A Document is just an entity class that holds a title, a text and an
identifier. The class doesn't actually do anything, just checks that its
constructor parameters are OK and holds its data. It acts as an in-memory
representation of a real document (probably stored in the database), to be
annotated, or already annotated, with its correspondent named entities.