Add Identifier value class
An Identifier is just a positive integer (represented via a Value Class) that
will be used in all Entity classes that need to have an ID. Having this class
frees all the Entity classes that need an ID of the responsibility of checking
if a given Integer (maybe passed in in its constructor) is a valid ID
(positive).

The one and only thing that Identifier does is hold a positive integer, and of
course validate that its constructor parameter is one of those.