Make Document use Identifier as ID and Sentence as title
Document is modified to be a final class (disallowing inheritance from it), to
use an Identifier value class as its ID instead of a plain Integer, and to use
a Sentece class as its title instead of a plain String. Using an Identifier for
ID frees Document from the responsibility of checking if a given value is a
positive Integer, and using a Sentence as Title frees it from checking if it is
empty and the future need of trimming white spaces in it (to cleanly store it
in the database).