Add SLF4J-Nop and ScalaMock dependencies
Slick uses SLF4J for its own debug logging, so an implementation is needed.
SLF4J-Nop provides an implementation that just does nothing, so it actually
disables logging from Slick. This is good actually, since it is not needed at
the moment. It will be replaced in the future by another logging framework if
its outputs were needed.

ScalaMock is a framework for creating Mock objects (for testing purposes) for
Scala. It has ScalaTest support, so that it can be both easily integrated.
Current ScalaMock does not have a compiler plugin like previous ones (for Scala
2.9 and older), so it is a bit more limited, but it will probably be enough.