Add Annotation entity class and tests
An Annotation represents an annotated named entity in a document, meaning a
sentence that has been recognized by a NER tool inside a document text. It
holds an identifier, the original sentence, that same sentence normalized (to
an standard representation, eg: "person", "boy", "human"... all could be
normalized to "homo sapiens"), a Category (as defined by the entity
Enumeration) that holds the type of the annotated sentence, a reference to the
document that was annotated with it and absolute start and ending positions of
the original sentence in the document text.