Add Virtuslab Unicorn dependency
Unicorn is a small library that enables writing type safe primary and foreign
keys in Slick applications. It is written to be used with Slick 1.x, support
for Slick 2.x is in development (Slick 2.0.0 has been released just a month
ago), so Slick dependency is downgraded also to 1.0.1 to accommodate.

Also added Play-Slick dependency, a library that makes Slick a first class
citizen inside the Play Framework, and established JDBC as a dependency to Play
Framework.