Add Play Framework config file
Application.conf inside "conf" directory establishes the configuration for Play
Framework. The only important stuff there is the configuration of the database:
default one has been set to a MySQL DB in localhost, and test is set to a H2
Database (in memory); and the final line that tells Slick where to look Scala
files that define tables to generate Play Evolution's SQL Files.