Remove Identifier and modify Document and Annotation models
Identifier class is no longer needed, since Unicorn library provides a type
safe BaseId and IdCompanion traits that can be used to define Identifiers for
each model that needs them. So it has been replaced by DocumentId and
AnnotationId. Annotation and Document classes are modified to be in sync with
the deletion of the Identifier class.

Defined Documents and Annotations SQL Tables via Slick Table definition. These
are subject to change since I think they are not properly defined: maybe split
Annotations table in two, one that relates Documents with Named Entities and
holds the original text and positions and another that defines the Named
Entities and holds just an ID and the original text.