Fix misspell in Slick conf and enable Ebean for models
Slick configuration in application.conf had a misspelling (smartdrugsearch was
written as smardrugsearch). Also enabled Ebean for model package, it is not
necessary right now, but if disabled a non-sense error is given on execution,
it's just enabled to make that error not appear (it does not affect anything
actually, just shuts out that there is an error in the logs).