Update configuration and evolutions
Evolutions have been disabled in the configuration file, since they are giving
me some problems, and the database table definitions are quite stable at this
point. The problem is in the table generation by Slick, that is just not quite
right at the moment, will solve it in the future if using evolutions becomes
necessary, but at this point it is simpler to just write by hand the SQL
scripts (which are added here, replacing the one generated by Slick).

Also moduralized a bit more the configuration files, placing the logging
settings into its own file, and updated some settings in the database files.