Remove Virtuslab Unicorn dependency
Virtuslab's Unicorn library is giving me some small headaches, so it is removed
and will implement a by-hand Identifier class that simulates what Unicorn
offers. Also, because of this library dropping, Slick and hence Play-Slick can
be upgraded to latests versions (that were not compatible with Unicorn).

Upgraded ScalaTest too to latest version just published (2.1.0) and ScalaMock
support for it.