Implement NERAdapter base Trait
Added a small implementation of a NERAdapter Trait that all NER Tool adapters
must conform to (and extend). It implements an Akka Actor that can receive
documents and annotate them storing its results in the database.

Implementations of this trait mus provide a DAL object as attribute (probably
received in its constructor), and its database Session should be established by
sending a Session message to the actor. They also should implement the
annotate(Document):Unit method providing the annotator logic required to parse
the document and store its NamedEntities and Annotations in the database as
expected.