Implement Linnaeus NER Adapter and tests
LinnaeusNERAdapter is a NERAdapter (Akka Actor) that performs annotations on
Documents, trying to find the species that its text holds. Linnaeus uses a set
of files (containing a dictionary of species) that allow it to recognize a
bunch of species and return its NCBI Taxonomy ID and position inside analyzed
text. That NCBI Taxonomy ID is used to recover the species' scientific name
from the NCBI Taxonomy[1] web service (part of eUtils[2]) and create a
NamedEntity object that will be stored in the database (if not previously
there). Once that NamedEntity has been stored, an Annotation object is created
holding the original text in document and its position, linking it via Foreign
Keys to the DocumentID and the previous NamedEntityID, and storing it in the
database.

Usage of this LinnaeusNERAdapter should be similar to this:

    val linnaeusActor = Akka.system.actorOf(Props(classOf[LinnaeusNERAdapter], myDALObject))
    linnaeusActor ! myDatabaseSession
    linnaeusActor ! myDocument

It should be noted that a database Session object should be sent to the
Linnaeus actor prior it does any work, since it needs to interact with a
database. That session should be of the same type of the JDBCProfile that the
DAL object that was passed in to the LinnaeusNERAdapter constructor (meaning:
type of `myDatabaseSession` should be the same as
`myDALObject.profile.simple.Session`).

**IMPORTANT**: This requirement is currently not forced by the implementation
(it will probably be in a nearby future), so sending a document without having
previously sent a session will just fail horrendously and abruptly.

[1] http://www.ncbi.nlm.nih.gov/taxonomy
[2] http://www.ncbi.nlm.nih.gov/books/NBK25501/