Increment wait time for WS response in LinnaeusNER
Response from the Taxonomy WebService of eUtils (NCBI) was expected to be
received within 5 seconds. This timeout was giving some random problems while
testing, as it looks that sometimes that response will be slower.

Waiting time has been incremented to 10 seconds. I think we can safely assume
that if within 10 seconds a response from the WebService has not been received,
it will be a good idea to just report a Failure, and let client code do
whatever it wants with that.