Implement Oscar NER Adapter
OscarNERAdapter is a NERAdapter (Akka Actor) that performs annotations on
documents, trying to find the chemical compounds that its text holds. Oscar
uses its own internal chemical name recognizer and normalizer, so its usage is
quite simple. Each chemical compound found is normalized into InChI and stored
into database, along with the original word or sentence that was used in the
document.

Usage of this OscarNERAdapter should be similar to this:

    val oscarActor = Akka.system.actorOf(Props[OscarNERAdapter])
    oscarActor ! myDatabaseSession
    oscarActor ! myDocument

Testing is a bit slow, since no different dictionary of chemical compounds than
the one that Oscar internally uses is established. It will probably be changed
to use a small subset of its big dictionary in testing mode, allowing a faster
response from the Actor.