Implement ABNER annotator adapter
ABNERAdapter is a NERAdapter (Akka Actor) that performs annotations on
documents, trying to find important keywords in them (proteins, dna, rna, cell
lines and cell types). It cannot normalize them to an standard representation,
so the one and only thing that it does to normalize is check if the keyword has
already been stored in the database, if not it just creates it (a real solution
to this would be to use something like Mora[1] instead).

Usage of this ABNERAdapter should be similar to this:

  val abnerActor = Akka.system.actorOf(Props[ABNERAdapter])
  abnerActor ! myDatabaseSession
  abnerActor ! myDocument

The Category entity has also been modified to hold more types of keywords
(necessary for ABNER types).

[1] http://moara.dacya.ucm.es/download.html