Database access refactor
Database access classes have been refactored. Now, DatabaseProfile holds a
Database reference, from where "withTransaction", "withConnection",
"withSession" and "createSession" can be called. This achieves a better design,
allowing easier testing and usage of the different DAOs. DAOs and Annotators
tests have been refactored to reflect this new DatabaseProfile behaviour, and
they are now a lot simpler and cleaner. Both use now the BeforeAndAfter trait
of ScalaTest in its superclasses to create a session to the Testing database,
and create and drop the database tables.

This new design of the database objects allows the NERAdapters to not need a
previous step of configuration where a Session is passed. They can now create
directly a session with the DatabaseProfile that they were already creating.
And because of that, an implicit session parameter is no longer required, since
a session can be directly created wherever it is needed.

Also, SQL evolutions have been updated and re-integrated intro Git and Akka
library has been downgraded to 2.2.4 since it looks that Play 2.2.x can't
properly work with Akka 2.3.x.