Add Annotator base actor
Annotator Akka Actor serves as a facade to all the different annotators
(NERAdapters) in the system. It just passes them the documents to annotate and
logs the results of it (if finished correctly or something failed). Now, with
this Actor, annotating a document with all annotators is as simple as:

    val annotator = system.actorOf(Props[Annotator])
    annotator ! myDocument

And that's it! All done in a non-blocking way, son client code can continue
doing whatever is required while the document is being annotated.

NOTE: It is a bit ugly to have all NERAdapters actors created "by hand",
because that requires modifying the Annotator actor whenever a new NERAdapter
is to be used.  Maybe the list of NERAdapter classes can be obtained from a
configuration file, it could be cleaner. Will try to do it that way in a
future.