EUtils WSDL to Scala with scalaxb
scalaxb[1] has been installed as a sbt plugin, in order to create the mapping
scala classes from the Entrez Utilites SOAP Service[2]. The WSDL file is
placed in "src/main/wsdl" and generated code will be in package "soap" once
scalaxb generates it (since it is managed code, it will reside in
"target/scala*/src_managed" and not in "src/main/scala/").

A EUtils object has been created inside utils package object to wrap all calls
to that SOAP Services and present them to the rest of the code in an cleaner
way. With that created, LinnaeusNERAdapter has been cleaned up of that ugly web
service calls, and instead just use that EUtils object.

[1] http://scalaxb.org/
[2] http://www.ncbi.nlm.nih.gov/books/NBK43082/