Database configuration in one file only
Database configuration has been placed into a single database.conf file,
allowing its parameters to be override by environment variables. This way, a
default in-memory H2 database is configured, but it can be changed to a MySQL
database (or any other, given that its JDBC connector is present) when
required.

Some changes to code are required to correctly support this, mostly in testing
specs. But this way a cleaner way to initialize DatabaseProfile is achieved.