Change EUtils object to a class
The EUtils object has been changed to be a class, since a better design is
achieved this way. Also, its constructor is private-hidden and it can only be
constructed via its companion object factory, so previous "EUtils" calls can be
replaced with just "EUtils()" and will work the same way.