Drop support for JDK6 in Travis
JDK6 is not supported by this project, since java.nio.file is being used, and
that was introduced in JDK7. That lack of support is causing, obviously, Travis
to fail its build when using openjdk6. So, just remove it.

Also, AnnotatorBaseSpec in annotator package tests has been removed and
replaced with ActorBaseSpec in the top-level package, that way we can reuse it
in some more Actors outside of the annotator package.