Add SearcherAdapter and implementations
A SearcherAdapter trait and implementation classes using the ABNER, Oscar and
Linnaeus services have been added into the searcher package. All of them
(implementations) perform a normalization of the search terms and just return a
set of keywords (by searching in the database) that match that normalized
terms. Also, a SimpleSearcher implementation has been added, that does not
perform any normalization, and just search for the Keywords in the original
texts of Annotations in the database.