Implement DocumentStatsService
DocumentStatsService is an Akka Actor that calculates statistics on (Document,
Keyword) pairs and stores that results into the DocumentStats table of the
database. It is configured in the Global Settings (see 'Global.scala') to perform
that statistics computation between some intervals (configured in
'conf/stats.conf').

Currently, statistics performed are based in counting the number of appearances
of a Keyword in a Document, and calculation of a ratio of appearance by
dividing that number with the total appearances of that same Keyword in all
documents.