Implement Searcher and Account entity
Searcher is the top-level searcher, it's responsibility is to communicate with
all enabled searchers in "conf/searchers.conf" to retrieve a Set of Keywords
which it will use to obtain the Documents that have any match with the received
search terms. It will also sort those documents based on the stats computed by
the DocumentStatsService actor, and stored into the DocumentStats database
table.

Also, an Account entity and an Accounts table have been added. These will store
the accounts of the system. We will only want to have administrator accounts
right now, so no worries about permissions or role-based authorization: if a
user is logged in, then it is an admin (no other possibility at the moment),
and it can do whatever it wants.