Add controllers to all supported actions
Implementation of Controllers to all supported actions to be performed by the
systems have been added. There is an Authorization trait that can filter
request based on the state of a token header (if it is set and that token UUID
is stored in Cache, then it is logged in; otherwise, it is not). There is also
a JavaScript Router controller that returns all routes in JavaScript, in order
to easily be used by the client application.

Also, actions on common errors have been changed to JSON responses instead of
Play's default HTML pages.