Add Blocked field to Documents
Document entity (and database table) now has a "blocked" field that will be set
to true when that document is currently being annotated. Otherwise, it will
always be false. This prevents some bad behaviours in the annotator when
multiple requests to annotate the same document are send simultaneously or
while one is already in process. It is required to improve this "block" thing,
since its implementation in the Annotator is not fully correct, but it's
enough for now.

All EntityList implementations have been placed inside the EntityList.scala
file, and the trait has been marked as sealed. Also, non-required
implementations (keywords, annotations) have been removed.

Documents now also parse input to remove all xml/html entities when reading
from user input JSON (Forms) and when downloaded via the Entrez Utilities
service. This cleaning is done via the Jsoup library.