Implement Admin account handling in Interface
Admin account handling (list, create, edit and delete) has been implemented
inside the Administration Panel in the web interface. Creation and edition of
accounts does not currently verifies password nor notifies the stored email. In
a future version of SmartDrugSearch, it will be a good idea to not allow other
administrators to edit account passwords nor create passwords when creating an
account: just send an email to the account either with a random generated
password or with a link to set one. Also, some security in passwords should be
enforced, such as requiring a minimum length.

Document listing has also been improved, and a new interface for user defined
documents (manual creation) has been implemented.