Disallow deletion of blocked documents
Documents marked as blocked are no longer deletable, since deleting them can
cause some conflicts in the ongoing annotation process. It is enforced only in
DocumentsController in the server-side. Future versions of SmartDrugSearch
should move the checking also into lower layers (such as *.database.Queries).
In the client-side, it is only enforced by disabling the "Delete" button if the
document is blocked, and it is quite fine in that way, blocking logic should be
performed always on server side.